fix: unable to use submodule dependency class in test, throws NoClassDefFoundError
See https://stackoverflow.com/questions/74832230/unable-to-use-submodule-dependency-class-in-test-throws-noclassdeffounderror/74839873

The reason why dependsOnMigration is unable to use Migration class in the test package is because of `spring-boot-maven-plugin` that was configured in Migration submodule. As Migration is a uber-jar, and compiled by `spring-boot-maven-plugin`, the way it package the jar is slightly different and that is why it is not possible to reference the class in dependsOnMigration test package.

The solution is to remove `spring-boot-maven-plugin` if migration submodule is not meant to be a spring-boot application OR to create two separate jar for different usage (see https://www.baeldung.com/spring-boot-dependency)